refactored more of lvl_script
