+++ src/creature_groups.c
@@ -783,6 +783,78 @@ void leader_find_positions_for_followers(struct Thing *leadtng)
         creature_follower_pos_add(leadtng, ifollow, &pos);
     }
 }
+
+/**
+ * Spawns new creature parties. Makes given amount of the parties.
+ * @param party The party to be spawned.
+ * @param plyr_idx Player to own the creatures within group.
+ * @param location Where the party will be spawned.
+ * @param copies_num Amount of copies to be spawned.
+ * @return Gives leader of last party spawned.
+ */
+struct Thing *script_process_new_party(struct Party *party, PlayerNumber plyr_idx, TbMapLocation location, long copies_num)
+{
+    struct Thing* leadtng = INVALID_THING;
+    for (long i = 0; i < copies_num; i++)
+    {
+        struct Thing* grptng = INVALID_THING;
+        for (long k = 0; k < party->members_num; k++)
+        {
+          if (k >= GROUP_MEMBERS_COUNT)
+          {
+              ERRORLOG("Party too big, %d is the limit",GROUP_MEMBERS_COUNT);
+              break;
+          }
+          struct PartyMember* member = &(party->members[k]);
+          struct Thing* thing = script_create_new_creature(plyr_idx, member->crtr_kind, location, member->carried_gold, member->crtr_level);
+          if (!thing_is_invalid(thing))
+          {
+              struct CreatureControl* cctrl = creature_control_get_from_thing(thing);
+              cctrl->party_objective = member->objectv;
+              cctrl->wait_to_turn = game.play_gameturn + member->countdown;
+              if (thing_is_invalid(grptng))
+              {
+                  // If it is the first creature - set it as only group member and leader
+                  // Inside the thing, we don't need to mark it in any way (two creatures are needed to form a real group)
+                  SYNCDBG(5,"First member %s index %d",thing_model_name(thing),(int)thing->index);
+                  leadtng = thing;
+                  grptng = thing;
+              } else
+              {
+                  struct Thing* bestng = get_best_creature_to_lead_group(grptng);
+                  struct CreatureControl* bestctrl = creature_control_get_from_thing(bestng);
+                  // If current leader wants to defend, and current unit has an objective, new unit will be group leader.
+                  if ((cctrl->party_objective != CHeroTsk_DefendParty) && (bestctrl->party_objective == CHeroTsk_DefendParty))
+                  {
+                      add_creature_to_group_as_leader(thing, grptng);
+                      leadtng = thing;
+                  } else
+                  // if best and current unit want to defend party, or neither do, the strongest will be leader
+                  if (((cctrl->party_objective == CHeroTsk_DefendParty) && (bestctrl->party_objective == CHeroTsk_DefendParty)) || ((cctrl->party_objective != CHeroTsk_DefendParty) && (bestctrl->party_objective != CHeroTsk_DefendParty)))
+                  {
+                      if ((cctrl->explevel > bestctrl->explevel) || ((cctrl->explevel == bestctrl->explevel) && (get_creature_thing_score(thing) > get_creature_thing_score(bestng))))
+                      {
+                          add_creature_to_group_as_leader(thing, grptng);
+                          leadtng = thing;
+                      }
+                      else
+                      // If it's weaker than the current leader, joind as a group
+                      {
+                          add_creature_to_group(thing, grptng);
+                      }
+                  }
+                  else
+                  // If it wants to defend, but the group leader has an objective, just add it to group
+                  {
+                      add_creature_to_group(thing, grptng);
+                  }
+              }
+          }
+        }
+    }
+    return leadtng;
+}
+
 /******************************************************************************/
 #ifdef __cplusplus
 }
