+++ src/lvl_script.c
@@ -33,7 +33,6 @@
 #include "config_crtrmodel.h"
 #include "config_effects.h"
 #include "config_lenses.h"
-#include "config_magic.h"
 #include "config_rules.h"
 #include "config_terrain.h"
 #include "config_trapdoor.h"
@@ -70,9 +69,12 @@
 #include "music_player.h"
 #include "custom_sprites.h"
 #include "console_cmd.h"
+#include "creature_states_hero.h"
+
 #include "lvl_script_commands.h"
 #include "lvl_script_lib.h"
 #include "lvl_script_statehandler.h"
+#include "lvl_script_conditions.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -108,95 +110,7 @@ const struct NamedCommand player_desc[] = {
   {NULL,               0},
 };
 
-const struct NamedCommand variable_desc[] = {
-    {"MONEY",                       SVar_MONEY},
-    {"GAME_TURN",                   SVar_GAME_TURN},
-    {"BREAK_IN",                    SVar_BREAK_IN},
-    //{"CREATURE_NUM",              SVar_CREATURE_NUM},
-    {"TOTAL_DIGGERS",               SVar_TOTAL_DIGGERS},
-    {"TOTAL_CREATURES",             SVar_TOTAL_CREATURES},
-    {"TOTAL_RESEARCH",              SVar_TOTAL_RESEARCH},
-    {"TOTAL_DOORS",                 SVar_TOTAL_DOORS},
-    {"TOTAL_AREA",                  SVar_TOTAL_AREA},
-    {"TOTAL_CREATURES_LEFT",        SVar_TOTAL_CREATURES_LEFT},
-    {"CREATURES_ANNOYED",           SVar_CREATURES_ANNOYED},
-    {"BATTLES_LOST",                SVar_BATTLES_LOST},
-    {"BATTLES_WON",                 SVar_BATTLES_WON},
-    {"ROOMS_DESTROYED",             SVar_ROOMS_DESTROYED},
-    {"SPELLS_STOLEN",               SVar_SPELLS_STOLEN},
-    {"TIMES_BROKEN_INTO",           SVar_TIMES_BROKEN_INTO},
-    {"GOLD_POTS_STOLEN",            SVar_GOLD_POTS_STOLEN},
-    {"HEART_HEALTH",                SVar_HEART_HEALTH},
-    {"GHOSTS_RAISED",               SVar_GHOSTS_RAISED},
-    {"SKELETONS_RAISED",            SVar_SKELETONS_RAISED},
-    {"VAMPIRES_RAISED",             SVar_VAMPIRES_RAISED},
-    {"CREATURES_CONVERTED",         SVar_CREATURES_CONVERTED},
-    {"EVIL_CREATURES_CONVERTED",    SVar_EVIL_CREATURES_CONVERTED},
-    {"GOOD_CREATURES_CONVERTED",    SVar_GOOD_CREATURES_CONVERTED},
-    {"TIMES_ANNOYED_CREATURE",      SVar_TIMES_ANNOYED_CREATURE},
-    {"TIMES_TORTURED_CREATURE",     SVar_TIMES_TORTURED_CREATURE},
-    {"TOTAL_DOORS_MANUFACTURED",    SVar_TOTAL_DOORS_MANUFACTURED},
-    {"TOTAL_TRAPS_MANUFACTURED",    SVar_TOTAL_TRAPS_MANUFACTURED},
-    {"TOTAL_MANUFACTURED",          SVar_TOTAL_MANUFACTURED},
-    {"TOTAL_TRAPS_USED",            SVar_TOTAL_TRAPS_USED},
-    {"TOTAL_DOORS_USED",            SVar_TOTAL_DOORS_USED},
-    {"KEEPERS_DESTROYED",           SVar_KEEPERS_DESTROYED},
-    {"CREATURES_SACRIFICED",        SVar_CREATURES_SACRIFICED},
-    {"CREATURES_FROM_SACRIFICE",    SVar_CREATURES_FROM_SACRIFICE},
-    {"TIMES_LEVELUP_CREATURE",      SVar_TIMES_LEVELUP_CREATURE},
-    {"TOTAL_SALARY",                SVar_TOTAL_SALARY},
-    {"CURRENT_SALARY",              SVar_CURRENT_SALARY},
-    //{"TIMER",                     SVar_TIMER},
-    {"DUNGEON_DESTROYED",           SVar_DUNGEON_DESTROYED},
-    {"TOTAL_GOLD_MINED",            SVar_TOTAL_GOLD_MINED},
-    //{"FLAG",                      SVar_FLAG},
-    //{"ROOM",                      SVar_ROOM_SLABS},
-    {"DOORS_DESTROYED",             SVar_DOORS_DESTROYED},
-    {"CREATURES_SCAVENGED_LOST",    SVar_CREATURES_SCAVENGED_LOST},
-    {"CREATURES_SCAVENGED_GAINED",  SVar_CREATURES_SCAVENGED_GAINED},
-    {"ALL_DUNGEONS_DESTROYED",      SVar_ALL_DUNGEONS_DESTROYED},
-    //{"DOOR",                      SVar_DOOR_NUM},
-    {"GOOD_CREATURES",              SVar_GOOD_CREATURES},
-    {"EVIL_CREATURES",              SVar_EVIL_CREATURES},
-    {"TRAPS_SOLD",                  SVar_TRAPS_SOLD},
-    {"DOORS_SOLD",                  SVar_DOORS_SOLD},
-    {"MANUFACTURED_SOLD",           SVar_MANUFACTURED_SOLD},
-    {"MANUFACTURE_GOLD",            SVar_MANUFACTURE_GOLD},
-    {"TOTAL_SCORE",                 SVar_TOTAL_SCORE},
-    {"BONUS_TIME",                  SVar_BONUS_TIME},
-    {NULL,                           0},
-};
 
-const struct NamedCommand dk1_variable_desc[] = {
-    {"MONEY",                       SVar_MONEY},
-    {"GAME_TURN",                   SVar_GAME_TURN},
-    {"BREAK_IN",                    SVar_BREAK_IN},
-    //{"CREATURE_NUM",                SVar_CREATURE_NUM},
-    {"TOTAL_IMPS",                  SVar_TOTAL_DIGGERS},
-    {"TOTAL_CREATURES",             SVar_CONTROLS_TOTAL_CREATURES},
-    {"TOTAL_RESEARCH",              SVar_TOTAL_RESEARCH},
-    {"TOTAL_DOORS",                 SVar_TOTAL_DOORS},
-    {"TOTAL_AREA",                  SVar_TOTAL_AREA},
-    {"TOTAL_CREATURES_LEFT",        SVar_TOTAL_CREATURES_LEFT},
-    {"CREATURES_ANNOYED",           SVar_CREATURES_ANNOYED},
-    {"BATTLES_LOST",                SVar_BATTLES_LOST},
-    {"BATTLES_WON",                 SVar_BATTLES_WON},
-    {"ROOMS_DESTROYED",             SVar_ROOMS_DESTROYED},
-    {"SPELLS_STOLEN",               SVar_SPELLS_STOLEN},
-    {"TIMES_BROKEN_INTO",           SVar_TIMES_BROKEN_INTO},
-    {"GOLD_POTS_STOLEN",            SVar_GOLD_POTS_STOLEN},
-    //{"TIMER",                     SVar_TIMER},
-    {"DUNGEON_DESTROYED",           SVar_DUNGEON_DESTROYED},
-    {"TOTAL_GOLD_MINED",            SVar_TOTAL_GOLD_MINED},
-    //{"FLAG",                      SVar_FLAG},
-    //{"ROOM",                      SVar_ROOM_SLABS},
-    {"DOORS_DESTROYED",             SVar_DOORS_DESTROYED},
-    {"CREATURES_SCAVENGED_LOST",    SVar_CREATURES_SCAVENGED_LOST},
-    {"CREATURES_SCAVENGED_GAINED",  SVar_CREATURES_SCAVENGED_GAINED},
-    {"ALL_DUNGEONS_DESTROYED",      SVar_ALL_DUNGEONS_DESTROYED},
-    //{"DOOR",                      SVar_DOOR_NUM},
-    {NULL,                           0},
-};
 
 const struct NamedCommand controls_variable_desc[] = {
     {"TOTAL_DIGGERS",               SVar_CONTROLS_TOTAL_DIGGERS},
@@ -1797,92 +1711,6 @@ static void process_fx_line(struct ScriptFxLine *fx_line)
     }
 }
 
-TbBool parse_get_varib(const char *varib_name, long *varib_id, long *varib_type)
-{
-    char c;
-    int len = 0;
-    char arg[MAX_TEXT_LENGTH];
-
-    if (level_file_version > 0)
-    {
-        *varib_type = get_id(variable_desc, varib_name);
-    } else
-    {
-        *varib_type = get_id(dk1_variable_desc, varib_name);
-    }
-    if (*varib_type == -1)
-      *varib_id = -1;
-    else
-      *varib_id = 0;
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(creature_desc, varib_name);
-      *varib_type = SVar_CREATURE_NUM;
-    }
-    //TODO: list of lambdas
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(room_desc, varib_name);
-      *varib_type = SVar_ROOM_SLABS;
-    }
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(timer_desc, varib_name);
-      *varib_type = SVar_TIMER;
-    }
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(flag_desc, varib_name);
-      *varib_type = SVar_FLAG;
-    }
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(door_desc, varib_name);
-      *varib_type = SVar_DOOR_NUM;
-    }
-    if (*varib_id == -1)
-    {
-        *varib_id = get_id(trap_desc, varib_name);
-        *varib_type = SVar_TRAP_NUM;
-    }
-    if (*varib_id == -1)
-    {
-      *varib_id = get_id(campaign_flag_desc, varib_name);
-      *varib_type = SVar_CAMPAIGN_FLAG;
-    }
-    if (*varib_id == -1)
-    {
-        if (2 == sscanf(varib_name, "BOX%ld_ACTIVATE%c", varib_id, &c) && (c == 'D'))
-        {
-            // activateD
-            *varib_type = SVar_BOX_ACTIVATED;
-        }
-        else
-        {
-            *varib_id = -1;
-        }
-        if (2 == sscanf(varib_name, "SACRIFICED[%n%[^]]%c", &len, arg, &c) && (c == ']'))
-        {
-            *varib_id = get_id(creature_desc, arg);
-            *varib_type = SVar_SACRIFICED;
-        }
-        if (2 == sscanf(varib_name, "REWARDED[%n%[^]]%c", &len, arg, &c) && (c == ']'))
-        {
-            *varib_id = get_id(creature_desc, arg);
-            *varib_type = SVar_REWARDED;
-        }
-    }
-    if (*varib_id == -1)
-    {
-      SCRPTERRLOG("Unknown variable name, '%s'", varib_name);
-      return false;
-    }
-    return true;
-}
-
-
-
-
 
 
 static void player_reveal_map_area(PlayerNumber plyr_idx, long x, long y, long w, long h)
@@ -3073,90 +2901,6 @@ TbBool get_coords_at_meta_action(struct Coord3d *pos, PlayerNumber target_plyr_i
     return true;
 }
 
-long send_tunneller_to_point(struct Thing *thing, struct Coord3d *pos)
-{
-    struct CreatureControl* cctrl = creature_control_get_from_thing(thing);
-    cctrl->party.target_plyr_idx = -1;
-    setup_person_tunnel_to_position(thing, pos->x.stl.num, pos->y.stl.num, 0);
-    thing->continue_state = CrSt_TunnellerDoingNothing;
-    return 1;
-}
-
-TbBool script_support_send_tunneller_to_action_point(struct Thing *thing, long apt_idx)
-{
-    SYNCDBG(7,"Starting");
-    struct ActionPoint* apt = action_point_get(apt_idx);
-    struct Coord3d pos;
-    if (action_point_exists(apt)) {
-        pos.x.val = apt->mappos.x.val;
-        pos.y.val = apt->mappos.y.val;
-    } else {
-        ERRORLOG("Attempt to send to non-existing action point %d",(int)apt_idx);
-        pos.x.val = subtile_coord_center(map_subtiles_x/2);
-        pos.y.val = subtile_coord_center(map_subtiles_y/2);
-    }
-    pos.z.val = subtile_coord(1,0);
-    send_tunneller_to_point(thing, &pos);
-    return true;
-}
-
-TbBool script_support_send_tunneller_to_dungeon(struct Thing *creatng, PlayerNumber plyr_idx)
-{
-    SYNCDBG(7,"Send %s to player %d",thing_model_name(creatng),(int)plyr_idx);
-    struct Thing* heartng = get_player_soul_container(plyr_idx);
-    TRACE_THING(heartng);
-    if (thing_is_invalid(heartng))
-    {
-        WARNLOG("Tried to send %s to player %d which has no heart", thing_model_name(creatng), (int)plyr_idx);
-        return false;
-    }
-    struct Coord3d pos;
-    if (!get_random_position_in_dungeon_for_creature(plyr_idx, CrWaS_WithinDungeon, creatng, &pos)) {
-        WARNLOG("Tried to send %s to player %d but can't find position", thing_model_name(creatng), (int)plyr_idx);
-        return send_tunneller_to_point_in_dungeon(creatng, plyr_idx, &heartng->mappos);
-    }
-    if (!send_tunneller_to_point_in_dungeon(creatng, plyr_idx, &pos)) {
-        WARNLOG("Tried to send %s to player %d but can't start the task", thing_model_name(creatng), (int)plyr_idx);
-        return false;
-    }
-    SYNCDBG(17,"Moving %s to (%d,%d)",thing_model_name(creatng),(int)pos.x.stl.num,(int)pos.y.stl.num);
-    return true;
-}
-
-TbBool script_support_send_tunneller_to_dungeon_heart(struct Thing *creatng, PlayerNumber plyr_idx)
-{
-    SYNCDBG(7,"Send %s to player %d",thing_model_name(creatng),(int)plyr_idx);
-    struct Thing* heartng = get_player_soul_container(plyr_idx);
-    TRACE_THING(heartng);
-    if (thing_is_invalid(heartng)) {
-        WARNLOG("Tried to send %s to player %d which has no heart", thing_model_name(creatng), (int)plyr_idx);
-        return false;
-    }
-    if (!send_tunneller_to_point_in_dungeon(creatng, plyr_idx, &heartng->mappos)) {
-        WARNLOG("Tried to send %s to player %d but can't start the task", thing_model_name(creatng), (int)plyr_idx);
-        return false;
-    }
-    SYNCDBG(17,"Moving %s to (%d,%d)",thing_model_name(creatng),(int)heartng->mappos.x.stl.num,(int)heartng->mappos.y.stl.num);
-    return true;
-}
-
-TbBool script_support_send_tunneller_to_appropriate_dungeon(struct Thing *creatng)
-{
-    SYNCDBG(7,"Starting");
-    //return _DK_script_support_send_tunneller_to_appropriate_dungeon(thing);
-    PlayerNumber plyr_idx;
-    struct Coord3d pos;
-    plyr_idx = get_best_dungeon_to_tunnel_to(creatng);
-    if (plyr_idx == -1) {
-        ERRORLOG("Could not find appropriate dungeon to send %s to",thing_model_name(creatng));
-        return false;
-    }
-    if (!get_random_position_in_dungeon_for_creature(plyr_idx, CrWaS_WithinDungeon, creatng, &pos)) {
-        WARNLOG("Tried to send %s to player %d but can't find position", thing_model_name(creatng), (int)plyr_idx);
-        return false;
-    }
-    return send_tunneller_to_point_in_dungeon(creatng, plyr_idx, &pos);
-}
 
 static struct Thing *script_create_creature_at_location(PlayerNumber plyr_idx, ThingModel crmodel, TbMapLocation location)
 {
@@ -3294,76 +3038,7 @@ struct Thing *script_process_new_tunneler(unsigned char plyr_idx, TbMapLocation 
     return creatng;
 }
 
-/**
- * Spawns new creature parties. Makes given amount of the parties.
- * @param party The party to be spawned.
- * @param plyr_idx Player to own the creatures within group.
- * @param location Where the party will be spawned.
- * @param copies_num Amount of copies to be spawned.
- * @return Gives leader of last party spawned.
- */
-struct Thing *script_process_new_party(struct Party *party, PlayerNumber plyr_idx, TbMapLocation location, long copies_num)
-{
-    struct Thing* leadtng = INVALID_THING;
-    for (long i = 0; i < copies_num; i++)
-    {
-        struct Thing* grptng = INVALID_THING;
-        for (long k = 0; k < party->members_num; k++)
-        {
-          if (k >= GROUP_MEMBERS_COUNT)
-          {
-              ERRORLOG("Party too big, %d is the limit",GROUP_MEMBERS_COUNT);
-              break;
-          }
-          struct PartyMember* member = &(party->members[k]);
-          struct Thing* thing = script_create_new_creature(plyr_idx, member->crtr_kind, location, member->carried_gold, member->crtr_level);
-          if (!thing_is_invalid(thing))
-          {
-              struct CreatureControl* cctrl = creature_control_get_from_thing(thing);
-              cctrl->party_objective = member->objectv;
-              cctrl->wait_to_turn = game.play_gameturn + member->countdown;
-              if (thing_is_invalid(grptng))
-              {
-                  // If it is the first creature - set it as only group member and leader
-                  // Inside the thing, we don't need to mark it in any way (two creatures are needed to form a real group)
-                  SYNCDBG(5,"First member %s index %d",thing_model_name(thing),(int)thing->index);
-                  leadtng = thing;
-                  grptng = thing;
-              } else
-              {
-                  struct Thing* bestng = get_best_creature_to_lead_group(grptng);
-                  struct CreatureControl* bestctrl = creature_control_get_from_thing(bestng);
-                  // If current leader wants to defend, and current unit has an objective, new unit will be group leader.
-                  if ((cctrl->party_objective != CHeroTsk_DefendParty) && (bestctrl->party_objective == CHeroTsk_DefendParty))
-                  {
-                      add_creature_to_group_as_leader(thing, grptng);
-                      leadtng = thing;
-                  } else
-                  // if best and current unit want to defend party, or neither do, the strongest will be leader
-                  if (((cctrl->party_objective == CHeroTsk_DefendParty) && (bestctrl->party_objective == CHeroTsk_DefendParty)) || ((cctrl->party_objective != CHeroTsk_DefendParty) && (bestctrl->party_objective != CHeroTsk_DefendParty)))
-                  {
-                      if ((cctrl->explevel > bestctrl->explevel) || ((cctrl->explevel == bestctrl->explevel) && (get_creature_thing_score(thing) > get_creature_thing_score(bestng))))
-                      {
-                          add_creature_to_group_as_leader(thing, grptng);
-                          leadtng = thing;
-                      }
-                      else
-                      // If it's weaker than the current leader, joind as a group
-                      {
-                          add_creature_to_group(thing, grptng);
-                      }
-                  }
-                  else
-                  // If it wants to defend, but the group leader has an objective, just add it to group
-                  {
-                      add_creature_to_group(thing, grptng);
-                  }
-              }
-          }
-        }
-    }
-    return leadtng;
-}
+
 
 struct Thing *script_create_new_creature(PlayerNumber plyr_idx, ThingModel crmodel, TbMapLocation location, long carried_gold, long crtr_level)
 {
@@ -3848,247 +3523,6 @@ TbBool action_point_activated_by_player(ActionPointId apt_idx, PlayerNumber plyr
     return ((i & (1 << plyr_idx)) != 0);
 }
 
-long get_condition_value(PlayerNumber plyr_idx, unsigned char valtype, unsigned char validx)
-{
-    SYNCDBG(10,"Checking condition %d for player %d",(int)valtype,(int)plyr_idx);
-    struct Dungeon* dungeon;
-    struct DungeonAdd* dungeonadd;
-    struct Thing* thing;
-    switch (valtype)
-    {
-    case SVar_MONEY:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_money_owned;
-    case SVar_GAME_TURN:
-        return game.play_gameturn;
-    case SVar_BREAK_IN:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->times_breached_dungeon;
-    case SVar_CREATURE_NUM:
-        return count_player_creatures_of_model(plyr_idx, validx);
-    case SVar_TOTAL_DIGGERS:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->num_active_diggers;
-    case SVar_TOTAL_CREATURES:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->num_active_creatrs;
-    case SVar_TOTAL_RESEARCH:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_research_points / 256;
-    case SVar_TOTAL_DOORS:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_doors;
-    case SVar_TOTAL_AREA:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_area;
-    case SVar_TOTAL_CREATURES_LEFT:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_creatures_left;
-    case SVar_CREATURES_ANNOYED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->creatures_annoyed;
-    case SVar_BATTLES_LOST:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->battles_lost;
-    case SVar_BATTLES_WON:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->battles_won;
-    case SVar_ROOMS_DESTROYED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->rooms_destroyed;
-    case SVar_SPELLS_STOLEN:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->spells_stolen;
-    case SVar_TIMES_BROKEN_INTO:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->times_broken_into;
-    case SVar_GHOSTS_RAISED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.ghosts_raised;
-    case SVar_SKELETONS_RAISED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.skeletons_raised;
-    case SVar_VAMPIRES_RAISED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.vamps_created;
-    case SVar_CREATURES_CONVERTED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.creatures_converted;
-    case SVar_TIMES_ANNOYED_CREATURE:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.lies_told;
-    case SVar_TOTAL_DOORS_MANUFACTURED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.manufactured_doors;
-    case SVar_TOTAL_TRAPS_MANUFACTURED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.manufactured_traps;
-    case SVar_TOTAL_MANUFACTURED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.manufactured_items;
-    case SVar_TOTAL_TRAPS_USED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.traps_used;
-    case SVar_TOTAL_DOORS_USED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.doors_used;
-    case SVar_KEEPERS_DESTROYED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.keepers_destroyed;
-    case SVar_TIMES_LEVELUP_CREATURE:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.creatures_trained;
-    case SVar_TIMES_TORTURED_CREATURE:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.creatures_tortured;
-    case SVar_CREATURES_SACRIFICED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.creatures_sacrificed;
-    case SVar_CREATURES_FROM_SACRIFICE:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.creatures_from_sacrifice;
-    case SVar_TOTAL_SALARY:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.salary_cost;
-    case SVar_CURRENT_SALARY:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->creatures_total_pay;
-    case SVar_GOLD_POTS_STOLEN:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->gold_pots_stolen;
-    case SVar_HEART_HEALTH:
-        thing = get_player_soul_container(plyr_idx);
-        if (thing_is_dungeon_heart(thing))
-        {
-            return thing->health;
-        }
-        return 0;
-    case SVar_TIMER:
-        dungeon = get_dungeon(plyr_idx);
-        if (dungeon->turn_timers[validx].state)
-          return game.play_gameturn - dungeon->turn_timers[validx].count;
-        else
-          return 0;
-    case SVar_DUNGEON_DESTROYED:
-        return !player_has_heart(plyr_idx);
-    case SVar_TOTAL_GOLD_MINED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->lvstats.gold_mined;
-    case SVar_FLAG:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->script_flags[validx];
-    case SVar_ROOM_SLABS:
-        return get_room_slabs_count(plyr_idx, validx);
-    case SVar_DOORS_DESTROYED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->doors_destroyed;
-    case SVar_CREATURES_SCAVENGED_LOST:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->creatures_scavenge_lost;
-    case SVar_CREATURES_SCAVENGED_GAINED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->creatures_scavenge_gain;
-    case SVar_AVAILABLE_MAGIC: // IF_AVAILABLE(MAGIC)
-        return is_power_available(plyr_idx, validx);
-    case SVar_AVAILABLE_TRAP: // IF_AVAILABLE(TRAP)
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->mnfct_info.trap_amount_stored[validx%gameadd.trapdoor_conf.trap_types_count]
-              + dungeonadd->mnfct_info.trap_amount_offmap[validx%gameadd.trapdoor_conf.trap_types_count];
-    case SVar_AVAILABLE_DOOR: // IF_AVAILABLE(DOOR)
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->mnfct_info.door_amount_stored[validx%gameadd.trapdoor_conf.door_types_count]
-              + dungeonadd->mnfct_info.door_amount_offmap[validx%gameadd.trapdoor_conf.door_types_count];
-    case SVar_AVAILABLE_ROOM: // IF_AVAILABLE(ROOM)
-        dungeon = get_dungeon(plyr_idx);
-        return (dungeon->room_buildable[validx%ROOM_TYPES_COUNT] & 1);
-    case SVar_AVAILABLE_CREATURE: // IF_AVAILABLE(CREATURE)
-        dungeon = get_dungeon(plyr_idx);
-        if (creature_will_generate_for_dungeon(dungeon, validx)) {
-            return min(game.pool.crtr_kind[validx%CREATURE_TYPES_COUNT],dungeon->max_creatures_attracted - (long)dungeon->num_active_creatrs);
-        }
-        return 0;
-    case SVar_SLAB_OWNER: //IF_SLAB_OWNER
-    {
-        long varib_id = get_slab_number(plyr_idx, validx);
-        struct SlabMap* slb = get_slabmap_direct(varib_id);
-        return slabmap_owner(slb);
-    }
-    case SVar_SLAB_TYPE: //IF_SLAB_TYPE
-    {
-        long varib_id = get_slab_number(plyr_idx, validx);
-        struct SlabMap* slb = get_slabmap_direct(varib_id);
-        return slb->kind;
-    }
-    case SVar_CONTROLS_CREATURE: // IF_CONTROLS(CREATURE)
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->owned_creatures_of_model[validx%CREATURE_TYPES_COUNT]
-          - count_player_list_creatures_of_model_matching_bool_filter(plyr_idx, validx, creature_is_kept_in_custody_by_enemy_or_dying);
-    case SVar_CONTROLS_TOTAL_CREATURES:// IF_CONTROLS(TOTAL_CREATURES)
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->num_active_creatrs - count_player_creatures_not_counting_to_total(plyr_idx);
-    case SVar_CONTROLS_TOTAL_DIGGERS:// IF_CONTROLS(TOTAL_DIGGERS)
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->num_active_diggers - count_player_diggers_not_counting_to_total(plyr_idx);
-    case SVar_ALL_DUNGEONS_DESTROYED:
-    {
-        struct PlayerInfo* player = get_player(plyr_idx);
-        return all_dungeons_destroyed(player);
-    }
-    case SVar_DOOR_NUM:
-        return count_player_deployed_doors_of_model(plyr_idx, validx);
-    case SVar_TRAP_NUM:
-        return count_player_deployed_traps_of_model(plyr_idx, validx);
-    case SVar_GOOD_CREATURES:
-        dungeon = get_dungeon(plyr_idx);
-        return count_creatures_in_dungeon_of_model_flags(dungeon, 0, CMF_IsEvil|CMF_IsSpectator|CMF_IsSpecDigger);
-    case SVar_EVIL_CREATURES:
-        dungeon = get_dungeon(plyr_idx);
-        return count_creatures_in_dungeon_of_model_flags(dungeon, CMF_IsEvil, CMF_IsSpectator|CMF_IsSpecDigger);
-    case SVar_CONTROLS_GOOD_CREATURES:
-        dungeon = get_dungeon(plyr_idx);
-        return count_creatures_in_dungeon_controlled_and_of_model_flags(dungeon, 0, CMF_IsEvil|CMF_IsSpectator|CMF_IsSpecDigger);
-    case SVar_CONTROLS_EVIL_CREATURES:
-        dungeon = get_dungeon(plyr_idx);
-        return count_creatures_in_dungeon_controlled_and_of_model_flags(dungeon, CMF_IsEvil, CMF_IsSpectator|CMF_IsSpecDigger);
-    case SVar_CAMPAIGN_FLAG:
-        return intralvl.campaign_flags[plyr_idx][validx];
-    case SVar_BOX_ACTIVATED:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->box_info.activated[validx];
-    case SVar_SACRIFICED:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->creature_sacrifice[validx];
-    case SVar_REWARDED:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->creature_awarded[validx];
-    case SVar_EVIL_CREATURES_CONVERTED:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->evil_creatures_converted;
-    case SVar_GOOD_CREATURES_CONVERTED:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->good_creatures_converted;
-    case SVar_TRAPS_SOLD:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->traps_sold;
-    case SVar_DOORS_SOLD:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->doors_sold;
-    case SVar_MANUFACTURED_SOLD:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->traps_sold + dungeonadd->doors_sold;
-    case SVar_MANUFACTURE_GOLD:
-        dungeonadd = get_dungeonadd(plyr_idx);
-        return dungeonadd->manufacture_gold;
-    case SVar_TOTAL_SCORE:
-        dungeon = get_dungeon(plyr_idx);
-        return dungeon->total_score;
-    case SVar_BONUS_TIME:
-        return (game.bonus_time - game.play_gameturn);
-    default:
-        break;
-    };
-    return 0;
-}
 
 TbBool get_condition_status(unsigned char opkind, long val1, long val2)
 {
