+++ src/lvl_script_conditions.c
@@ -1,0 +1,377 @@
+/******************************************************************************/
+// Free implementation of Bullfrog's Dungeon Keeper strategy game.
+/******************************************************************************/
+/** @file lvl_script.c
+ *     Level script commands support.
+ * @par Purpose:
+ *     Load, recognize and maintain the level script.
+ * @par Comment:
+ *     None.
+ * @author   Tomasz Lis
+ * @date     12 Feb 2009 - 11 Apr 2014
+ * @par  Copying and copyrights:
+ *     This program is free software; you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation; either version 2 of the License, or
+ *     (at your option) any later version.
+ */
+/******************************************************************************/
+
+
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/******************************************************************************/
+#include "lvl_script_conditions.h"
+
+#include "globals.h"
+#include "dungeon_data.h"
+#include "config_magic.h"
+#include "game_legacy.h"
+#include "room_entrance.h"
+#include "creature_states.h"
+#include "keeperfx.hpp"
+
+ const struct NamedCommand variable_desc[] = {
+    {"MONEY",                       SVar_MONEY},
+    {"GAME_TURN",                   SVar_GAME_TURN},
+    {"BREAK_IN",                    SVar_BREAK_IN},
+    //{"CREATURE_NUM",              SVar_CREATURE_NUM},
+    {"TOTAL_DIGGERS",               SVar_TOTAL_DIGGERS},
+    {"TOTAL_CREATURES",             SVar_TOTAL_CREATURES},
+    {"TOTAL_RESEARCH",              SVar_TOTAL_RESEARCH},
+    {"TOTAL_DOORS",                 SVar_TOTAL_DOORS},
+    {"TOTAL_AREA",                  SVar_TOTAL_AREA},
+    {"TOTAL_CREATURES_LEFT",        SVar_TOTAL_CREATURES_LEFT},
+    {"CREATURES_ANNOYED",           SVar_CREATURES_ANNOYED},
+    {"BATTLES_LOST",                SVar_BATTLES_LOST},
+    {"BATTLES_WON",                 SVar_BATTLES_WON},
+    {"ROOMS_DESTROYED",             SVar_ROOMS_DESTROYED},
+    {"SPELLS_STOLEN",               SVar_SPELLS_STOLEN},
+    {"TIMES_BROKEN_INTO",           SVar_TIMES_BROKEN_INTO},
+    {"GOLD_POTS_STOLEN",            SVar_GOLD_POTS_STOLEN},
+    {"HEART_HEALTH",                SVar_HEART_HEALTH},
+    {"GHOSTS_RAISED",               SVar_GHOSTS_RAISED},
+    {"SKELETONS_RAISED",            SVar_SKELETONS_RAISED},
+    {"VAMPIRES_RAISED",             SVar_VAMPIRES_RAISED},
+    {"CREATURES_CONVERTED",         SVar_CREATURES_CONVERTED},
+    {"EVIL_CREATURES_CONVERTED",    SVar_EVIL_CREATURES_CONVERTED},
+    {"GOOD_CREATURES_CONVERTED",    SVar_GOOD_CREATURES_CONVERTED},
+    {"TIMES_ANNOYED_CREATURE",      SVar_TIMES_ANNOYED_CREATURE},
+    {"TIMES_TORTURED_CREATURE",     SVar_TIMES_TORTURED_CREATURE},
+    {"TOTAL_DOORS_MANUFACTURED",    SVar_TOTAL_DOORS_MANUFACTURED},
+    {"TOTAL_TRAPS_MANUFACTURED",    SVar_TOTAL_TRAPS_MANUFACTURED},
+    {"TOTAL_MANUFACTURED",          SVar_TOTAL_MANUFACTURED},
+    {"TOTAL_TRAPS_USED",            SVar_TOTAL_TRAPS_USED},
+    {"TOTAL_DOORS_USED",            SVar_TOTAL_DOORS_USED},
+    {"KEEPERS_DESTROYED",           SVar_KEEPERS_DESTROYED},
+    {"CREATURES_SACRIFICED",        SVar_CREATURES_SACRIFICED},
+    {"CREATURES_FROM_SACRIFICE",    SVar_CREATURES_FROM_SACRIFICE},
+    {"TIMES_LEVELUP_CREATURE",      SVar_TIMES_LEVELUP_CREATURE},
+    {"TOTAL_SALARY",                SVar_TOTAL_SALARY},
+    {"CURRENT_SALARY",              SVar_CURRENT_SALARY},
+    //{"TIMER",                     SVar_TIMER},
+    {"DUNGEON_DESTROYED",           SVar_DUNGEON_DESTROYED},
+    {"TOTAL_GOLD_MINED",            SVar_TOTAL_GOLD_MINED},
+    //{"FLAG",                      SVar_FLAG},
+    //{"ROOM",                      SVar_ROOM_SLABS},
+    {"DOORS_DESTROYED",             SVar_DOORS_DESTROYED},
+    {"CREATURES_SCAVENGED_LOST",    SVar_CREATURES_SCAVENGED_LOST},
+    {"CREATURES_SCAVENGED_GAINED",  SVar_CREATURES_SCAVENGED_GAINED},
+    {"ALL_DUNGEONS_DESTROYED",      SVar_ALL_DUNGEONS_DESTROYED},
+    //{"DOOR",                      SVar_DOOR_NUM},
+    {"GOOD_CREATURES",              SVar_GOOD_CREATURES},
+    {"EVIL_CREATURES",              SVar_EVIL_CREATURES},
+    {"TRAPS_SOLD",                  SVar_TRAPS_SOLD},
+    {"DOORS_SOLD",                  SVar_DOORS_SOLD},
+    {"MANUFACTURED_SOLD",           SVar_MANUFACTURED_SOLD},
+    {"MANUFACTURE_GOLD",            SVar_MANUFACTURE_GOLD},
+    {"TOTAL_SCORE",                 SVar_TOTAL_SCORE},
+    {"BONUS_TIME",                  SVar_BONUS_TIME},
+    {NULL,                           0},
+};
+
+
+const struct NamedCommand dk1_variable_desc[] = {
+    {"MONEY",                       SVar_MONEY},
+    {"GAME_TURN",                   SVar_GAME_TURN},
+    {"BREAK_IN",                    SVar_BREAK_IN},
+    //{"CREATURE_NUM",                SVar_CREATURE_NUM},
+    {"TOTAL_IMPS",                  SVar_TOTAL_DIGGERS},
+    {"TOTAL_CREATURES",             SVar_CONTROLS_TOTAL_CREATURES},
+    {"TOTAL_RESEARCH",              SVar_TOTAL_RESEARCH},
+    {"TOTAL_DOORS",                 SVar_TOTAL_DOORS},
+    {"TOTAL_AREA",                  SVar_TOTAL_AREA},
+    {"TOTAL_CREATURES_LEFT",        SVar_TOTAL_CREATURES_LEFT},
+    {"CREATURES_ANNOYED",           SVar_CREATURES_ANNOYED},
+    {"BATTLES_LOST",                SVar_BATTLES_LOST},
+    {"BATTLES_WON",                 SVar_BATTLES_WON},
+    {"ROOMS_DESTROYED",             SVar_ROOMS_DESTROYED},
+    {"SPELLS_STOLEN",               SVar_SPELLS_STOLEN},
+    {"TIMES_BROKEN_INTO",           SVar_TIMES_BROKEN_INTO},
+    {"GOLD_POTS_STOLEN",            SVar_GOLD_POTS_STOLEN},
+    //{"TIMER",                     SVar_TIMER},
+    {"DUNGEON_DESTROYED",           SVar_DUNGEON_DESTROYED},
+    {"TOTAL_GOLD_MINED",            SVar_TOTAL_GOLD_MINED},
+    //{"FLAG",                      SVar_FLAG},
+    //{"ROOM",                      SVar_ROOM_SLABS},
+    {"DOORS_DESTROYED",             SVar_DOORS_DESTROYED},
+    {"CREATURES_SCAVENGED_LOST",    SVar_CREATURES_SCAVENGED_LOST},
+    {"CREATURES_SCAVENGED_GAINED",  SVar_CREATURES_SCAVENGED_GAINED},
+    {"ALL_DUNGEONS_DESTROYED",      SVar_ALL_DUNGEONS_DESTROYED},
+    //{"DOOR",                      SVar_DOOR_NUM},
+    {NULL,                           0},
+};
+
+
+long get_condition_value(PlayerNumber plyr_idx, unsigned char valtype, unsigned char validx)
+{
+    SYNCDBG(10,"Checking condition %d for player %d",(int)valtype,(int)plyr_idx);
+    struct Dungeon* dungeon;
+    struct DungeonAdd* dungeonadd;
+    struct Thing* thing;
+    switch (valtype)
+    {
+    case SVar_MONEY:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_money_owned;
+    case SVar_GAME_TURN:
+        return game.play_gameturn;
+    case SVar_BREAK_IN:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->times_breached_dungeon;
+    case SVar_CREATURE_NUM:
+        return count_player_creatures_of_model(plyr_idx, validx);
+    case SVar_TOTAL_DIGGERS:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->num_active_diggers;
+    case SVar_TOTAL_CREATURES:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->num_active_creatrs;
+    case SVar_TOTAL_RESEARCH:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_research_points / 256;
+    case SVar_TOTAL_DOORS:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_doors;
+    case SVar_TOTAL_AREA:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_area;
+    case SVar_TOTAL_CREATURES_LEFT:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_creatures_left;
+    case SVar_CREATURES_ANNOYED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->creatures_annoyed;
+    case SVar_BATTLES_LOST:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->battles_lost;
+    case SVar_BATTLES_WON:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->battles_won;
+    case SVar_ROOMS_DESTROYED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->rooms_destroyed;
+    case SVar_SPELLS_STOLEN:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->spells_stolen;
+    case SVar_TIMES_BROKEN_INTO:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->times_broken_into;
+    case SVar_GHOSTS_RAISED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.ghosts_raised;
+    case SVar_SKELETONS_RAISED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.skeletons_raised;
+    case SVar_VAMPIRES_RAISED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.vamps_created;
+    case SVar_CREATURES_CONVERTED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.creatures_converted;
+    case SVar_TIMES_ANNOYED_CREATURE:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.lies_told;
+    case SVar_TOTAL_DOORS_MANUFACTURED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.manufactured_doors;
+    case SVar_TOTAL_TRAPS_MANUFACTURED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.manufactured_traps;
+    case SVar_TOTAL_MANUFACTURED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.manufactured_items;
+    case SVar_TOTAL_TRAPS_USED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.traps_used;
+    case SVar_TOTAL_DOORS_USED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.doors_used;
+    case SVar_KEEPERS_DESTROYED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.keepers_destroyed;
+    case SVar_TIMES_LEVELUP_CREATURE:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.creatures_trained;
+    case SVar_TIMES_TORTURED_CREATURE:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.creatures_tortured;
+    case SVar_CREATURES_SACRIFICED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.creatures_sacrificed;
+    case SVar_CREATURES_FROM_SACRIFICE:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.creatures_from_sacrifice;
+    case SVar_TOTAL_SALARY:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.salary_cost;
+    case SVar_CURRENT_SALARY:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->creatures_total_pay;
+    case SVar_GOLD_POTS_STOLEN:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->gold_pots_stolen;
+    case SVar_HEART_HEALTH:
+        thing = get_player_soul_container(plyr_idx);
+        if (thing_is_dungeon_heart(thing))
+        {
+            return thing->health;
+        }
+        return 0;
+    case SVar_TIMER:
+        dungeon = get_dungeon(plyr_idx);
+        if (dungeon->turn_timers[validx].state)
+          return game.play_gameturn - dungeon->turn_timers[validx].count;
+        else
+          return 0;
+    case SVar_DUNGEON_DESTROYED:
+        return !player_has_heart(plyr_idx);
+    case SVar_TOTAL_GOLD_MINED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->lvstats.gold_mined;
+    case SVar_FLAG:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->script_flags[validx];
+    case SVar_ROOM_SLABS:
+        return get_room_slabs_count(plyr_idx, validx);
+    case SVar_DOORS_DESTROYED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->doors_destroyed;
+    case SVar_CREATURES_SCAVENGED_LOST:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->creatures_scavenge_lost;
+    case SVar_CREATURES_SCAVENGED_GAINED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->creatures_scavenge_gain;
+    case SVar_AVAILABLE_MAGIC: // IF_AVAILABLE(MAGIC)
+        return is_power_available(plyr_idx, validx);
+    case SVar_AVAILABLE_TRAP: // IF_AVAILABLE(TRAP)
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->mnfct_info.trap_amount_stored[validx%gameadd.trapdoor_conf.trap_types_count]
+              + dungeonadd->mnfct_info.trap_amount_offmap[validx%gameadd.trapdoor_conf.trap_types_count];
+    case SVar_AVAILABLE_DOOR: // IF_AVAILABLE(DOOR)
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->mnfct_info.door_amount_stored[validx%gameadd.trapdoor_conf.door_types_count]
+              + dungeonadd->mnfct_info.door_amount_offmap[validx%gameadd.trapdoor_conf.door_types_count];
+    case SVar_AVAILABLE_ROOM: // IF_AVAILABLE(ROOM)
+        dungeon = get_dungeon(plyr_idx);
+        return (dungeon->room_buildable[validx%ROOM_TYPES_COUNT] & 1);
+    case SVar_AVAILABLE_CREATURE: // IF_AVAILABLE(CREATURE)
+        dungeon = get_dungeon(plyr_idx);
+        if (creature_will_generate_for_dungeon(dungeon, validx)) {
+            return min(game.pool.crtr_kind[validx%CREATURE_TYPES_COUNT],dungeon->max_creatures_attracted - (long)dungeon->num_active_creatrs);
+        }
+        return 0;
+    case SVar_SLAB_OWNER: //IF_SLAB_OWNER
+    {
+        long varib_id = get_slab_number(plyr_idx, validx);
+        struct SlabMap* slb = get_slabmap_direct(varib_id);
+        return slabmap_owner(slb);
+    }
+    case SVar_SLAB_TYPE: //IF_SLAB_TYPE
+    {
+        long varib_id = get_slab_number(plyr_idx, validx);
+        struct SlabMap* slb = get_slabmap_direct(varib_id);
+        return slb->kind;
+    }
+    case SVar_CONTROLS_CREATURE: // IF_CONTROLS(CREATURE)
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->owned_creatures_of_model[validx%CREATURE_TYPES_COUNT]
+          - count_player_list_creatures_of_model_matching_bool_filter(plyr_idx, validx, creature_is_kept_in_custody_by_enemy_or_dying);
+    case SVar_CONTROLS_TOTAL_CREATURES:// IF_CONTROLS(TOTAL_CREATURES)
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->num_active_creatrs - count_player_creatures_not_counting_to_total(plyr_idx);
+    case SVar_CONTROLS_TOTAL_DIGGERS:// IF_CONTROLS(TOTAL_DIGGERS)
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->num_active_diggers - count_player_diggers_not_counting_to_total(plyr_idx);
+    case SVar_ALL_DUNGEONS_DESTROYED:
+    {
+        struct PlayerInfo* player = get_player(plyr_idx);
+        return all_dungeons_destroyed(player);
+    }
+    case SVar_DOOR_NUM:
+        return count_player_deployed_doors_of_model(plyr_idx, validx);
+    case SVar_TRAP_NUM:
+        return count_player_deployed_traps_of_model(plyr_idx, validx);
+    case SVar_GOOD_CREATURES:
+        dungeon = get_dungeon(plyr_idx);
+        return count_creatures_in_dungeon_of_model_flags(dungeon, 0, CMF_IsEvil|CMF_IsSpectator|CMF_IsSpecDigger);
+    case SVar_EVIL_CREATURES:
+        dungeon = get_dungeon(plyr_idx);
+        return count_creatures_in_dungeon_of_model_flags(dungeon, CMF_IsEvil, CMF_IsSpectator|CMF_IsSpecDigger);
+    case SVar_CONTROLS_GOOD_CREATURES:
+        dungeon = get_dungeon(plyr_idx);
+        return count_creatures_in_dungeon_controlled_and_of_model_flags(dungeon, 0, CMF_IsEvil|CMF_IsSpectator|CMF_IsSpecDigger);
+    case SVar_CONTROLS_EVIL_CREATURES:
+        dungeon = get_dungeon(plyr_idx);
+        return count_creatures_in_dungeon_controlled_and_of_model_flags(dungeon, CMF_IsEvil, CMF_IsSpectator|CMF_IsSpecDigger);
+    case SVar_CAMPAIGN_FLAG:
+        return intralvl.campaign_flags[plyr_idx][validx];
+    case SVar_BOX_ACTIVATED:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->box_info.activated[validx];
+    case SVar_SACRIFICED:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->creature_sacrifice[validx];
+    case SVar_REWARDED:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->creature_awarded[validx];
+    case SVar_EVIL_CREATURES_CONVERTED:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->evil_creatures_converted;
+    case SVar_GOOD_CREATURES_CONVERTED:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->good_creatures_converted;
+    case SVar_TRAPS_SOLD:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->traps_sold;
+    case SVar_DOORS_SOLD:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->doors_sold;
+    case SVar_MANUFACTURED_SOLD:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->traps_sold + dungeonadd->doors_sold;
+    case SVar_MANUFACTURE_GOLD:
+        dungeonadd = get_dungeonadd(plyr_idx);
+        return dungeonadd->manufacture_gold;
+    case SVar_TOTAL_SCORE:
+        dungeon = get_dungeon(plyr_idx);
+        return dungeon->total_score;
+    case SVar_BONUS_TIME:
+        return (game.bonus_time - game.play_gameturn);
+    default:
+        break;
+    };
+    return 0;
+}
+
+
+
+/******************************************************************************/
+#ifdef __cplusplus
+}
+#endif
